pyproject.toml
[tool.poetry]
name = "repl_python3_FrolferBot"
version = "0.1.0"
description = ""
authors = ["jpjordangit <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
pendulum = "^2.1.2"
python-dotenv = "^0.19.2"
Flask = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
